+++ src/spiderswitch/index/__init__.py
@@ -5,6 +5,7 @@ from .builder import ModelCapabilityIndex, ModelIndexEntry, build_index
 from .experience import ExperienceStore, SubjectiveRecord
 from .schema import StructuredCapabilities
 from .service import ModelIndexService
+from .store import default_index_path, load_index, save_index
 
 __all__ = [
     "ExperienceStore",
@@ -14,4 +15,7 @@ __all__ = [
     "StructuredCapabilities",
     "SubjectiveRecord",
     "build_index",
+    "default_index_path",
+    "load_index",
+    "save_index",
 ]